CoreRoot/requirements.txt
Django==5.2.1
psycopg2-binary==2.9.10
djangorestframework==3.16.0
django-filter==25.1

